Allow multiple RegEx matches per line
Edited the format for the headerConversionList so that RegEx can be used to pull specific information and match/post it to differing columns, rather than taking the entire content of the match and posting it to a single column.

Also now doesn't move the pdf files unless Airtable confirms the data was imported correctly.