Transition toward OOP design
Just learning about OOP, so this could be better. Would have been much better if I started the project with these ideas in mind. As-is, it takes the old code and attempts to jam it into an OOP design. Lots of work needs to be done to make it easier to read and work with, but for now, it works and uses 100 less lines of code.