Reload conversionlists.py added, change appendToDebugLog argument handling
Now reloads the conversionlists.py file when new files are found in the folder. This allows the script to continue running while changes are made to conversionlists.py in the background. Any errors are logged in the Debug log.

appendToDebugLog now takes a general error message along with keyword arguments as messages to append to the Debug log, making it easier to adapt to various situations.