Add initialization, suspend file processing
Will now create the folders if they do not exist. Then checks if the syntax is correct in conversionlists.py. If not, it will not process files, move files in the main folder to the Suspended folder, and append to the debug log. If it is correct, it will take any files out of the Suspended folder and place them in the main folder to be processed.

Fixed appendToDebugLog, now places error message in the log.