Create inputs class, add folder for holding input classes. Move PDF processing functions to that folder.
New inputs are to be added to a .py file in the inputs folder. They should be imported into main.py inside the inputs class. They should be passed the inventoryQueue and errorQueue objects so they can send instances of the inventoryObject class to the main internal database and send errors to the main error log.